Remove gson as dependancy
We use the internal GSON library.

Similarly, this is the final coremod patch for non-development environments. Currently, the mod will load in non-development environments, however the rendered player model does not render properly. This is one of the final issues before I prepare the mod for release.

The change adds a message indicating if the mod is disabled if no patches are applied & links to a support website