Add a menu to copy a skin from other players in the world #24
This release adds a new UI to "copy" the skin from another player in your world. The list of players is scrollable (using the mouse) in case it goes of the screen. Some UI unused UI elements were removed and a couple of translations were added. Additionally, the animations were split up between the players model and the cape model, so the cape movement now works when the player model animation is disabled. These animations are no longer frame dependant, and will work across different framerates by immitating the vanilla function of precomputing animations in the client update (tick) events. Some unused UI elements were also removed to combat the gradual increase in size of the mod.