Adds the new translation system and old buttons #29
This release replaces the built-in minecraft translation system with a dedicated JSON-based translation system. Similarly, a lot of fat has been trimmed from unused methods and classes. The installer has also been tweaked to ask if you'd like to delete mod files, and if the user agrees it will scan jar files to check if they are the mod instead of just using the name of the file. An attempt has also been made to tweak optifine.