Begins rework of cache system ( #6 ) and adds updater
This release includes a new bleeding edge cache system (currently a bit less reliable than the previous version).
It also includes an implementation of Apagoge to support GitHub releases (the mod will now tell you when there is an update).

I've fixed the issue with the "old buttons" text not being translated, and added two new configurable options as well
1. To disable the updater if you don't want it
2. To disable the cache clearing itself on startup

More files were moved around in this release. This is likely the final move for them.